Organized it in one main function
